+++ kernel/src/platform/aarch64/mod.rs
@@ -1,0 +1,3 @@
+//! Platforms using the AArch64 architecture.
+
+pub mod qemu_virt;